refactor to algorithms
