+++ src/lib.rs
@@ -3,5 +3,6 @@ mod sinks;
 mod abi;
 mod pb;
 mod utils;
+mod algorithms;
 mod debug;
 mod debug_store;
