+++ src/maps.rs
@@ -1,12 +1,13 @@
 use crate::abi::{self};
+use crate::algorithms::{compute_keccak_address_map, find_erc20_balance_changes_algorithm1, find_erc20_balance_changes_algorithm2, get_all_child_calls};
 use crate::pb::erc20::types::v1::{BalanceChange, BalanceChangeType, Events, Transfer};
-use crate::utils::{clock_to_date, compute_keccak_address_map, get_all_child_calls, get_owner_from_keccak_address_map, index_to_version, is_erc20_valid_address, is_erc20_valid_balance, StorageKeyToAddressMap};
+use crate::utils::{clock_to_date, index_to_version};
 use abi::erc20::events::Transfer as TransferAbi;
 use substreams::errors::Error;
-use substreams::log::info;
+
+use substreams::pb::substreams::Clock;
 use substreams::scalar::BigInt;
 use substreams::Hex;
-use substreams::pb::substreams::Clock;
 use substreams_ethereum::pb::eth::v2::{Block, Call, Log, StorageChange, TransactionTrace, TransactionTraceStatus};
 use substreams_ethereum::Event;
 
@@ -147,120 +148,5 @@ pub fn iter_balance_changes(transfers: Vec<(TransactionTrace, Call, Log, Transfe
             out.push((trx.clone(), call.clone(), log.clone(), transfer.clone(), owner, storage_changes, BalanceChangeType::BalanceChangeType2));
         }
     }
-    out
-}
-
-// algorithm #1 (normal case)
-fn find_erc20_balance_changes_algorithm1(
-    call: &Call,
-    transfer: &TransferAbi,
-    keccak_address_map: &StorageKeyToAddressMap,
-) -> Vec<(Vec<u8>, StorageChange)> {
-    let mut out = Vec::new();
-
-    for storage_change in &call.storage_changes {
-        // Check if the transfer matches the storage change balance changes
-        if !is_erc20_valid_balance(transfer, storage_change) {
-            continue;
-        }
-
-        // extract the owner address
-        let owner = match get_owner_from_keccak_address_map(keccak_address_map, &storage_change) {
-            Some(address) => address,
-            None => continue
-        };
-
-        // make sure owner is either the sender or receiver
-        if !is_erc20_valid_address(&owner, transfer) {
-            info!("owner={} does not match transfer from={} to={}", Hex(owner), Hex(&transfer.from), Hex(&transfer.to));
-            continue;
-        }
-        out.push((owner, storage_change.clone()));
-    }
-    out
-}
-
-// algorithm #2 (case where storage changes are not in the same call as the transfer event)
-fn find_erc20_balance_changes_algorithm2(
-    child_calls: Vec<Call>,
-    transfer: &TransferAbi,
-    keccak_address_map: &StorageKeyToAddressMap,
-) -> Vec<(Vec<u8>, StorageChange)> {
-    let mut out = Vec::new();
-
-    //get all storage changes for these calls:
-    let mut storage_changes = Vec::new();
-    for call in child_calls.iter() {
-        storage_changes.extend(call.storage_changes.clone());
-    }
-
-    let mut total_sent = BigInt::zero();
-    let mut total_received = BigInt::zero();
-
-    //check if any of the storage changes match the transfer.to or transfer.from
-    for storage_change in storage_changes.clone().iter() {
-        let owner = match get_owner_from_keccak_address_map(keccak_address_map, &storage_change) {
-            Some(address) => address,
-            None => continue
-        };
-
-        if !is_erc20_valid_address(&owner, transfer) {
-            info!("owner={} does not match transfer from={} to={}", Hex(owner), Hex(&transfer.from), Hex(&transfer.to));
-            continue;
-        }
-
-        let old_balance = BigInt::from_unsigned_bytes_be(&storage_change.old_value);
-        let new_balance = BigInt::from_unsigned_bytes_be(&storage_change.new_value);
-
-        let balance_change = new_balance - old_balance;
-        if balance_change < BigInt::zero() {
-            total_sent = total_sent + balance_change.neg();
-        } else {
-            total_received = total_received + balance_change;
-        };
-
-        out.push((owner, storage_change.clone()));
-    }
-
-    if total_sent == transfer.value {
-        return out;
-    }
-
-    let mut diff = total_sent - total_received;
-    if diff < BigInt::zero() {
-        diff = diff.neg();
-    }
-
-    // look for a storage change that matches the diff
-    for storage_change in storage_changes.iter() {
-
-        // Check if the transfer matches the storage change balance changes
-        let owner = match get_owner_from_keccak_address_map(keccak_address_map, &storage_change) {
-            Some(address) => address,
-            None => continue
-        };
-
-        // make sure owner is either the sender or receiver
-        if !is_erc20_valid_address(&owner, transfer) {
-            info!("owner={} does not match transfer from={} to={}", Hex(owner), Hex(&transfer.from), Hex(&transfer.to));
-            continue;
-        }
-
-        let old_balance = BigInt::from_unsigned_bytes_be(&storage_change.old_value);
-        let new_balance = BigInt::from_unsigned_bytes_be(&storage_change.new_value);
-
-        let mut balance_change = new_balance - old_balance;
-        if balance_change < BigInt::zero() {
-            balance_change = balance_change.neg();
-        }
-
-        if balance_change != diff {
-            info!("Algo2: Balance change does not match transfer value. Balance change: {}, transfer value: {}", balance_change, transfer.value);
-            continue;
-        }
-
-        out.push((owner, storage_change.clone()));
-    }
-
     out
 }
